Created directories for suse_9_x86_64 and suse_10_x86_64. Populated them with links to dotkits in linux_x86_64. Needed to create these in trunk before merging trunk changes into development branch. Dotkit upgrade coming soon to a shell near you.
Note: New dotkits built in least-common-denominator mode should 
still go to linux_x86_64, with symlinks added to directories for
supported OS releases. As far as I can tell there is no way to 
have a "fallback" location for dotkits since the OS specific 
directories corresponding to $SYS_TYPE appear to be the path of 
last resort.




git-svn-id: https://subversion01.broadinstitute.org/APSG@429 95e82132-2061-11dd-9cc4-33af95911ecc