Make .prompt++ simpler and hopefully more robust.
Add .load_prompt++ dotkit, a dynamic prompt for bash that 
sets the prompt to

user@hostname pwd $ 

and sets the color for "hostname" based on system load. 
Yes, load is a useless statistic, but dynamically changing 
prompt color seemed like a neat trick. Screenshot at
http://www/~jbh/load_prompt.png showing the result of hitting
enter as load climbs to 4 then returns to < .50




git-svn-id: https://subversion01.broadinstitute.org/APSG@478 95e82132-2061-11dd-9cc4-33af95911ecc